Arduino and ESP 32 Codes
Here the Arduino handle the motion of the robot by allowing the robot to move by following a black color line. The ESP 32 handles the stopping the robot and object picking mechanism according to the user need. It allows the robot to bring the object just by selecting which object that he/she wants to bring. The whole system functionality is based on the precious readings of set of sensors.